Feat: cleaned up marker package
