+++ packages/markers/models_test.go
@@ -2,10 +2,9 @@ package markers
 
 import (
 	"fmt"
-	"testing"
-
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"testing"
 )
 
 func TestMarker(t *testing.T) {
@@ -232,4 +231,23 @@ func TestReferencingMarkers(t *testing.T) {
 	assert.Equal(t, NewMarkers(), unmarshaledReferencingMarkers.Get(13))
 
 	fmt.Println(unmarshaledReferencingMarkers)
+}
+
+func TestSequence(t *testing.T) {
+	sequence := NewSequence(1337, NewMarkers(
+		&Marker{1, 3},
+		&Marker{2, 6},
+	), 7)
+
+	assert.Equal(t, SequenceID(1337), sequence.ID())
+	assert.Equal(t, uint64(7), sequence.Rank())
+	assert.Equal(t, Index(7), sequence.HighestIndex())
+
+	marshaledSequence := sequence.Bytes()
+	unmarshaledSequence, consumedBytes, err := SequenceFromBytes(marshaledSequence)
+	require.NoError(t, err)
+	assert.Equal(t, len(marshaledSequence), consumedBytes)
+	assert.Equal(t, sequence.ID(), unmarshaledSequence.ID())
+	assert.Equal(t, sequence.Rank(), unmarshaledSequence.Rank())
+	assert.Equal(t, sequence.HighestIndex(), unmarshaledSequence.HighestIndex())
 }
